Add: added furumaru about for top page
