+++ web/user/src/locales/en_us.json
@@ -103,6 +103,7 @@
       "pastMarcheLinkText": "View Past Live-Streams",
       "viewMoreText": "View More",
       "productsLinkText": "View Products for Sale",
+      "product": "Product List",
       "archiveListSubTitle": "videos",
       "productListSubTitle": "product videos",
       "experienceListSubTitle": "experience videos",
